refactor: rename shared to UI for clarity
