+++ src/app/join-us/JoinUsBox.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Link from 'next/link';
 import Image from 'next/image';
-import Button from '@/shared/Button/Button';
+import Button from '@/ui/Button/Button';
 
 interface JoinUsBoxProps {
 	role: string;
